# build/swagger.yaml
openapi: 3.0.3
info:
  title: Swagger BookStore
  description: Book Store RestFullAPi
  version: 1.0.11
servers:
  - url: http://localhost:3000/v1/api
...
paths:
...
  /auth/refreshToken:
    post:
      tags:
        - Auth
      operationId: Refreshtoken
      description: Refreshtoken
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc0FkbWluIjoiZmFsc2UiLCJpYXQiOjE3MDE4NTcxNTYsImV4cCI6MTcwMTg1NzE4Nn0.J0ldtd3_o5224q65AWZkgYjvfhO_37O2R4EG0XxOL0c
                idUser:
                  example: 2
        required: true
      responses:
        '200':
          description: refreshtoken success
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MiwiaXNBZG1pbiI6ImZhbHNlIiwiaWF0IjoxNzAyMTk0MDQ0LCJleHAiOjE3MDIxOTQwNzR9.N3_Ikq9Acqxoq0ypAh43GE9Zmb26_bZwiLpftqTZ9qE

        '403':
          description: Refresh token is not valid
...
  /user/{id}:
    get:
      tags:
        - 'Users'
      operationId: GetUser
      parameters:
        - name: id
          in: path
          description: ID user
          required: true
          schema:
            type: integer

      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: integer
                    description: User ID
                  username:
                    type: string
                    description: Username of the user
                  email:
                    type: string
                    format: email
                    description: Email address of the user
                  firstName:
                    type: string
                    description: First name of the user
                  lastName:
                    type: string
                    description: Last name of the user
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      name: Token
      in: header
      description: "Provide your API key with the 'Bearer' prefix"